+++ history.md
@@ -1,6 +1,6 @@
 # v1.0.33 - 2016/10/27
 
-* Added `#addServicePermissions` to `key` sub-module
+* Added `#grantClientAccess` to `key` sub-module
 
 # v1.0.32 - 2016/09/21